requirements-airflow.txt
# LangGraph and LangChain
langgraph
langchain
langchain-community
langchain-openai

# LangFuse for observability
langfuse

# OpenAI
openai>=1.0.0

# Database
psycopg2-binary==2.9.9
redis==5.0.0

# HTTP client
httpx==0.25.0
aiohttp==3.9.0

# Data validation
pydantic==2.0.0

# Utils
python-dotenv==1.0.0
tenacity==8.2.3

# For agent tools
pillow==10.0.0  # Image processing
pandas==2.0.0  # Data processing
numpy==1.24.0  # Numerical operations

# Additional Airflow providers if needed
apache-airflow-providers-celery==3.5.0